requirements.txt
djangorestframework==3.14.0
python-dotenv==1.0.0
mysqlclient==2.2.0
gunicorn==20.1.0
celery[redis]
pytest-django==4.7.0